Add /lib/triage-as*.swc to .gitignore
These files are only used by flashbuilder. The maven build will pull them out from the maven repository instead.
It seems better not to include these files in the source repository to avoid confusion with the lates libraries in the maven repository.
Developers that would like to include these files in the flashbuilder lib folder should do that manually.